@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -32,6 +32,9 @@ dependencies:
   sliding_up_panel: ^2.0.0+1
   # Added this to easily get page-transitions
   page_transition: "^2.0.4"
+  # Added this to get a better DropdownButton
+  dropdown_button2: ^1.1.1
+
   
 dev_dependencies:
   flutter_test:
